Release updates and template code organization
1. Update changelogs with 0.1.0 release information.
2. Update package.json's to 0.1.0 (docs-preview remains at current version).
3. Break up template logic to allow for better flow control.
4. Add a step to clean up existing template files prior to running download.  Uses a timeout to wait before showing quickpick.  Will likely utilize async/await in the future.
