+++ docs-article-templates/src/helper/github.ts
@@ -3,15 +3,14 @@
 import * as fs from "fs";
 import os = require("os");
 import * as path from "path";
-import * as vscode from "vscode";
 import { output } from "../extension";
 import * as common from "./common";
 
-const docsAuthoringDirectory = path.join(os.homedir(), "Docs Authoring");
-export const templateDirectory = path.join(docsAuthoringDirectory, "Templates");
+export const docsAuthoringDirectory = path.join(os.homedir(), "Docs Authoring");
+export const templateDirectory = path.join(docsAuthoringDirectory, "templates");
 
 // download a copy of the template repo to the "docs authoring" directory.  no .git-related files will be generated by this process.
-export function downloadRepo() {
+export async function downloadRepo() {
     const download = require("download-git-repo");
     const templateRepo = "MicrosoftDocs/content-templates";
     download(templateRepo, docsAuthoringDirectory, (err) => {
@@ -23,10 +22,18 @@ export function downloadRepo() {
 }
 
 // the download process is on a repo-level so this function will be used to delete any files pulled down by the download process.
-export function cleanupDownloadFiles() {
-    fs.readdir(docsAuthoringDirectory, (err, files) => {
+export async function cleanupDownloadFiles(templates?: boolean) {
+    let workingDirectory: string;
+
+    if (templates) {
+        workingDirectory = templateDirectory;
+    } else {
+        workingDirectory = docsAuthoringDirectory;
+    }
+
+    fs.readdir(workingDirectory, (err, files) => {
         files.forEach((file) => {
-            const fullFilePath = path.join(docsAuthoringDirectory, file);
+            const fullFilePath = path.join(workingDirectory, file);
             fs.stat(path.join(fullFilePath), (error, stats) => {
                 if (stats.isFile()) {
                     fs.unlinkSync(fullFilePath);
