+++ docs-authoring-pack/CHANGELOG.md
@@ -1,13 +1,26 @@
 # Change Log
 
+## 0.1.0 (June 18th, 2018)
+
+Added the following extensions to authoring pack:
+
+- docsmsft.docs-article-templates
+- streetsidesoftware.code-spell-checker
+
+Replaced ms-docfx.DocFX extension with docsmsft.docs-preview.
+
 ## 0.0.3 (May 10th, 2018)
-- Update readme 
+
+- Update readme
 
 ## 0.0.2 (April 30th, 2018)
+
 - Documentation updates
 
 ## 0.0.1 (April 3rd, 2018)
+
 First release that includes the following extensions:
+
 - DavidAnson.vscode-markdownlint
 - ms-docfx.DocFX
 - docsmsft.docs-markdown
