+++ docs-authoring-pack/package.json
@@ -3,7 +3,7 @@
   "displayName": "Docs Authoring Pack",
   "description": "Collection of extensions to assist with content development for docs.microsoft.com",
   "icon": "images/docs-logo-ms.png",
-  "version": "0.0.2",
+  "version": "0.1.0",
   "publisher": "docsmsft",
   "engines": {
     "vscode": "^1.20.0"
@@ -23,8 +23,10 @@
   ],
   "extensionDependencies": [
     "DavidAnson.vscode-markdownlint",
-    "ms-docfx.DocFX",
-    "docsmsft.docs-markdown"
+    "docsmsft.docs-preview",
+    "docsmsft.docs-markdown",
+    "docsmsft.docs-article-templates",
+    "streetsidesoftware.code-spell-checker"
   ],
   "license": "MIT"
 }
